backend/requirements.txt
Flask
ultralytics
Pillow
requests
python-multipart
google-generativeai
